requirements_versions.txt
torchsde==0.2.6
einops==0.10.0
transformers==5.3.0
diffusers==0.37.0
safetensors==0.6.1
accelerate==0.24.0
pyyaml==7.0
pillow==10.5.0
scipy==1.14.0
tqdm==4.66.4
psutil==6.0.0
pytorch_lightning==2.3.3
omegaconf==2.4.0
gradio==4.8.0
pygit2==1.15.1
opencv-contrib-python-headless==4.10.0.84
httpx==0.27.0
onnxruntime==1.18.1
timm==1.0.7
numpy==1.26.4
tokenizers==0.19.1
packaging==26.0
rembg==2.0.57
groundingdino-py==0.4.0
segment_anything==1.0